pyproject.toml
[project]
name = "cloudflare-ddns"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "load-dotenv>=0.1.0",
    "requests>=2.32.4",
]
